+++ src/main/kotlin/dev/pawfect/addons/config/features/EmojisConfig.kt
@@ -1,0 +1,9 @@
+package dev.pawfect.addons.config.features
+
+import com.google.gson.annotations.Expose
+
+class EmojisConfig {
+
+    @Expose
+    var enabled: Boolean = true
+}